integrated with order success

+++ src/main/java/com/fikadu/payment/entity/Payment.java
@@ -19,10 +19,10 @@ public class Payment {
     @GeneratedValue
     private Long id;
     private Long userId;
-    private Long orderId;
-    private Long totalCost;
-    private Long dasherPrice;
-    private Long restaurantPrice;
+    private long orderId;
+    private Double totalCost;
+    private Double dasherPrice;
+    private Double restaurantPrice;
     private String email;
     private String userName;
     private String restaurantEmail;
